IPD/ai-rag-medical/requirements.txt
fastapi==0.115.6
starlette==0.41.3
uvicorn==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
mcp
chromadb>=0.6.0
sentence-transformers>=3.0.0
